% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchors.R
\name{anchors}
\alias{anchors}
\title{Defines a new \code{anchor} object}
\usage{
anchors(predicates = logical(0))
}
\arguments{
\item{predicates}{a vector of \code{predicate} objects}
}
\value{
A new \code{anchors} object containing the properties \code{predicates}
}
\description{
Defines a new \code{anchor} object
}
\examples{
# Create predicates and combine them into an anchor.
pred_1 <- predicate(
  feature = "x",
  operator = `>`,
  constant = 0.8
)

pred_2 <- predicate(
  feature = "y",
  operator = `<`,
  constant = 0.9
)

anchor <- anchors(
  predicates = c(pred_1, pred_2)
)

anchor
}
