% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour_vis.R
\name{bounding_box}
\alias{bounding_box}
\alias{bounding_boxes}
\title{S7 class to hold the bounding box created by \code{make_anchors}}
\usage{
bounding_box(bounds_tbl, target_inst_row, point_colors)

bounding_boxes(boxes = class_missing)
}
\arguments{
\item{bounds_tbl}{Tibble containing two rows and (p+1) columns of p predictors}

\item{target_inst_row}{Tibble of one row and p columns for the target instance}

\item{point_colors}{Vector of 2^p length or 1 value}

\item{boxes}{A vector of S7 objects of class bounding box}
}
\value{
S7 object of class bounding box

S7 object of class bounding boxes
}
\description{
S7 class to hold the bounding box created by \code{make_anchors}

S7 class to hold a list of bounding boxes
}
