% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mab_utils.R
\name{create_anchor_inst}
\alias{create_anchor_inst}
\title{Creates an instance of the anchor class}
\usage{
create_anchor_inst(bounds, interest_cols)
}
\arguments{
\item{bounds}{The upper and lower bounds of each column of interest}

\item{interest_cols}{the columns of interest}
}
\value{
an instance of the anchor class
}
\description{
Creates an instance of the anchor class
}
