% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{envir_to_bounds}
\alias{envir_to_bounds}
\title{Lookup function to get value of upper and lower bounds for the current state}
\usage{
envir_to_bounds(current_envir, envir, interest_cols)
}
\arguments{
\item{current_envir}{List of indexes, the current state}

\item{envir}{The current environment}

\item{interest_cols}{Columns of interest}
}
\value{
A tibble of 1 x (2*p) where p is the number of columns of interest
}
\description{
The current state of the multi armed bandit is marked based on the indices
in the list of each combination of column and lower and upper bound type.
}
