% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend.R
\name{extend}
\alias{extend}
\alias{extend,kultarr::anchors-method}
\title{Extend an anchor}
\usage{
extend(x, pred)

## S7 method for class <kultarr::anchors>
extend(x, pred)
}
\arguments{
\item{x}{Object of S7 class \code{anchors}.}

\item{pred}{Object of S7 class \code{predicate}.}
}
\value{
An \code{anchors} object with the additional predicate.
}
\description{
\code{extend()} adds a new predicate to an anchor's existing set of
predicates, returning an updated \code{anchors} object.
}
\section{Methods}{
\code{extend()} is an S7 generic with methods available for the following
classes:

\itemize{
\item \code{kultarr::anchors}
}
}

\examples{
pred_1 <- predicate(
  feature = "x",
  operator = `>`,
  constant = 0.8
)

pred_2 <- predicate(
  feature = "y",
  operator = `<`,
  constant = 0.9
)

anchor <- anchors(
  predicates = c(pred_1)
)

# Add another predicate to the anchor.
extended_anchor <- extend(anchor, pred_2)

extended_anchor

}
