% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mab_utils.R
\name{generate_cutpoints}
\alias{generate_cutpoints}
\title{Generates lower and upper bounds around a given instance}
\usage{
generate_cutpoints(dataset, instance_id, interest_columns)
}
\arguments{
\item{dataset}{Cutpoints will be generated from in between the points in the dataset}

\item{instance_id}{The index of the instance of interest}

\item{interest_coluns}{The columns of \code{dataset} to consider when creating lower and upper bounds}
}
\value{
A list. Contains lower and upper bounds for each specific column of interest.
}
\description{
Generates lower and upper bounds around a given instance
}
