% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_perturbations}
\alias{generate_perturbations}
\title{Generate Perturbations Around a Point of Interest}
\usage{
generate_perturbations(
  data,
  instance,
  interest_columns,
  radius = 0.1,
  step = 0.01
)
}
\arguments{
\item{data}{Training data frame}

\item{instance}{Row number of point of interest}

\item{interest_columns}{The columns of the dataset used for generating anchors}

\item{radius}{Perturbation radius (default: 0.1)}

\item{step}{Step size for perturbations (default: 0.01)}
}
\value{
A data frame of perturbed points
}
\description{
Generate Perturbations Around a Point of Interest
}
