% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{envir_to_bounds}
\alias{envir_to_bounds}
\alias{update_bounds}
\alias{create_anchor_inst}
\alias{generate_environment}
\title{Lookup function to get value of upper and lower bounds for the current state}
\usage{
envir_to_bounds(current_envir, envir, interest_cols)

update_bounds(current_envir, envir, actions, selected_action)

create_anchor_inst(bounds, interest_cols)

generate_environment(dataset, instance_id, interest_columns, bin_edges)
}
\arguments{
\item{current_envir}{List of indexes, the current state}

\item{envir}{The current environment}

\item{interest_cols}{the columns of interest}

\item{selected_action}{The index of selected action from the actions list}

\item{bounds}{The upper and lower bounds of each column of interest}

\item{dataset}{Cutpoints will be generated from in between the points in the dataset}

\item{instance_id}{The index of the instance of interest}

\item{bin_edges}{Output of \code{define_bin_edges} function}

\item{interest_coluns}{The columns of \code{dataset} to consider when creating lower and upper bounds}
}
\value{
A tibble of 1 x (2*p) where p is the number of columns of interest

New bounds

an instance of the anchor class

A list. Contains lower and upper bounds for each specific column of interest.
}
\description{
The current state of the multi armed bandit is marked based on the indices
in the list of each combination of column and lower and upper bound type.
}
