% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mab.R
\name{run_mab}
\alias{run_mab}
\title{Run Multi-Armed bandit algorithm}
\usage{
run_mab(
  n_games,
  n_epochs,
  dataset,
  instance_id,
  environment,
  interest_cols,
  dist_func,
  model_func,
  class_ind,
  seed = 145,
  verbose
)
}
\arguments{
\item{n_games}{Numeric. Number of games to play}

\item{n_epochs}{Numeric. Number of epochs in a single game}

\item{dataset}{The dataset to run algorithm on}

\item{instance_id}{The index of the target observation in the datasert}

\item{environment}{The environment of poss}

\item{interest_cols}{The columns of interest}

\item{dist_func}{Function that takes n as an argument and returns
a data.frame of size n x p where p is the number of variables in the dataset}

\item{model_func}{Function that takes in a data.frame of size m x p
where m is the number of rows and p is the number of variables in the dataset
and returns a vector of predictions of size m x 1}

\item{class_ind}{Numeric. The index of the required class when ordered alphabetically}

\item{seed}{Numeric. Seed to be used for the Multi-Armed Bandit algorithm.
This ensures that the results stay consistent}

\item{verbose}{Logical. Whether to print out diagnostics of the Multi-Armed Bandit Algorithm}
}
\value{
A tibble of 1 x (2*p) where p is the number of columns of interest
}
\description{
Run Multi-Armed bandit algorithm
}
