% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mab.R
\name{select_action}
\alias{select_action}
\title{Function to decide on the appropriate actions}
\usage{
select_action(n_actions, success_probs, failure_probs)
}
\arguments{
\item{n_actions}{Number of possible actions}

\item{success_probs}{Numeric vector with length equal to n_actions containing values between 0-1}

\item{failure_probs}{Numeric vector with length equal to n_actions containing values between 0-1}
}
\value{
Numeric value in the range of 1 to n_actions
}
\description{
Function to decide on the appropriate actions
}
