% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mab_utils.R
\name{update_bounds}
\alias{update_bounds}
\title{Advance the current environment}
\usage{
update_bounds(current_envir, envir, actions, selected_action)
}
\arguments{
\item{current_envir}{List of indexes, the current state}

\item{selected_action}{The index of selected action from the actions list}
}
\value{
New bounds
}
\description{
Advance the current environment
}
