% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mab_utils.R
\name{validate_environment}
\alias{validate_environment}
\title{Function to validate whether
every edge case of the environment can contain the row}
\usage{
validate_environment(e, instance_row)
}
\description{
Function to validate whether
every edge case of the environment can contain the row
}
