% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour_vis.R
\name{anchor_tour}
\alias{anchor_tour}
\title{Data structure to hold information necessary to animate tours}
\usage{
anchor_tour(b_boxes, data, point_colors)
}
\arguments{
\item{b_boxes}{S7 object of type boundary boxes}

\item{data}{data.frame of points to visualize on tours}

\item{point_colors}{Vector of one or nrow(data)}
}
\description{
Data structure to hold information necessary to animate tours
}
