% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchors.R
\name{anchors}
\alias{anchors}
\title{Defines a new \code{anchor} object}
\usage{
anchors(predicates = logical(0))
}
\arguments{
\item{predicates}{a vector of \code{predicate} objects}
}
\value{
A new \code{anchors} object containing the properties \code{predicates}
}
\description{
Defines a new \code{anchor} object
}
