% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchors.R, R/coverage.R, R/extend.R,
%   R/precision.R, R/predicate.R, R/satisfies.R
\name{anchors}
\alias{anchors}
\alias{coverage}
\alias{extend}
\alias{precision}
\alias{predicate}
\alias{satisfies}
\title{S7 class to define anchors}
\usage{
anchors(predicates = logical(0))

coverage(x, ...)

extend(x, ...)

precision(x, ...)

predicate(
  feature = character(0),
  operator = function() NULL,
  constant = integer(0)
)

satisfies(x, ...)
}
\arguments{
\item{predicates}{a vector of objects with the S7 class of predicate}

\item{x}{Object of S7 class anchor}

\item{feature}{Character. Defines the column that the predicate is applied to}

\item{operator}{Binary operator that works with the column and the constant given}

\item{constant}{Value to compare \code{feature} with using \code{operator}. Can be numeric/character/logical}

\item{pred}{Object of S7 class predicate}

\item{model}{a predict function that will provide the predicted labels given a dataset}

\item{dist}{the function that can be used to generate samples by providing an argument n. Should return a dataframe with proper column names.}

\item{n_samples}{the number of samples to generate from \code{dist} (the perturbation distribution)}

\item{data}{The dataframe to apply anchors on. Can be one instance or an entire dataset.}
}
\value{
Object of S7 class anchors

Numeric. Coverage of anchor

named vector of proportions

A logical vector indicating whether the anchors satisfies \code{data}
}
\description{
S7 class to define anchors

Generic function calculate coverage of an anchor

Generic function to extend anchors

Generic function to calculate precision of anchors

S7 class to define a predicate

Generic function to calculate the number of rows in a dataset satisfied by an anchor
}
