% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour_vis.R
\name{animate_anchor}
\alias{animate_anchor}
\title{Generic function to visualize tours}
\usage{
animate_anchor(x, ...)
}
\arguments{
\item{x}{An object of type anchor_tour}

\item{...}{Additional arguments passed to display_xy}

\item{gif_file}{The file location to save the gif file}

\item{tour_path}{An object of type 'tour_path' from the tourr package. Defaults to grand_tour()}

\item{width}{the width of the gif file. Defaults to 500}

\item{height}{the height of the gif file. Defaults to 500}

\item{frames}{the number of frames to be included in the gif file. Defaults to 360}

\item{loop}{Logical. Defaults to TRUE}
}
\value{
None. Saves GIF at file location
}
\description{
Generic function to visualize tours
}
