% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage}
\alias{coverage}
\title{Generic function calculate coverage of an anchor}
\usage{
coverage(x, ...)
}
\arguments{
\item{samples}{the dataset to test coverage on}
}
\value{
Numeric. Coverage of anchor
}
\description{
The coverage of an anchor is defined as average number of observations satisfied by an anchor
}
