% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage}
\alias{coverage}
\alias{coverage,kultarr::anchors-method}
\title{Generic function calculate coverage of an anchor}
\usage{
coverage(x, samples)

## S7 method for class <kultarr::anchors>
coverage(x, samples)
}
\arguments{
\item{x}{anchors object}

\item{samples}{the dataset to test coverage on}
}
\value{
Numeric. Coverage of anchor
}
\description{
The coverage of an anchor is defined as average number of observations satisfied by an anchor
}
\section{Methods}{
\code{coverage()} is an S7 generic with methods available for the following classes:

\itemize{
\item \code{kultarr::anchors}
}
}

