% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mab_utils.R
\name{envir_to_bounds}
\alias{envir_to_bounds}
\title{Lookup function to get value of upper and lower bounds for the current state}
\usage{
envir_to_bounds(current_envir, envir, interest_cols)
}
\description{
The current state of the multi armed bandit is marked based on the indices
in the list of each combination of column and lower and upper bound type.
}
