% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend.R
\name{extend}
\alias{extend}
\title{Generic function to extend anchors}
\arguments{
\item{x}{Object of S7 class anchors}

\item{pred}{Object of S7 class predicate}
}
\value{
An anchor object with the additional predicate
}
\description{
Generic function to extend anchors
}
