% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mab.R
\name{get_reward}
\alias{get_reward}
\title{Reward function for Multi Armed Bandit}
\usage{
get_reward(
  new_anchor,
  round,
  dist_func,
  model_func,
  dataset,
  instance_id,
  class_ind = 1,
  verbose
)
}
\arguments{
\item{verbose}{Logical. Whether to print out diagnostics of the Multi-Armed Bandit Algorithm}
}
\description{
Reward function for Multi Armed Bandit
}
