% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mab.R
\name{make_anchors}
\alias{make_anchors}
\title{Make anchors}
\usage{
make_anchors(
  model,
  dataset,
  cols,
  instance,
  model_func,
  class_col,
  n_perturb_samples = 10000,
  n_games = 20,
  n_epochs = 100,
  seed = 145,
  verbose = FALSE
)
}
\arguments{
\item{model}{The model to be interrogated}

\item{dataset}{Dataset to use containing predictors and response variables.}

\item{cols}{Columns of interest}

\item{instance}{Id of the instance of interest in the training dataset}

\item{model_func}{Function that gives takes in any data and the model to give predictions}

\item{class_col}{Name of factor column containing class of interest}

\item{n_perturb_samples}{number of samples to be taken from the pertubation distribution}

\item{seed}{Numeric. Seed to be used for the Multi-Armed Bandit algorithm.
This ensures that the results stay consistent}

\item{verbose}{Logical. Whether to print out diagnostics of the Multi-Armed Bandit Algorithm}
}
\description{
Make anchors
}
