% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mab_utils.R
\name{make_perturb_distn}
\alias{make_perturb_distn}
\title{Make perturbation distribution function}
\usage{
make_perturb_distn(n, interest_cols, dataset, instance_id)
}
\arguments{
\item{n}{}

\item{instance_id}{}
}
\value{
A purrr partial function that takes in N and returns N number of sample points around the instance
}
\description{
Make perturbation distribution function
}
