% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision.R
\name{precision}
\alias{precision}
\alias{precision,any-method}
\alias{precision,kultarr::anchors-method}
\title{Precision of an anchor}
\usage{
precision(x, model, samples)

## S7 method for class <any>
precision(x, model, samples)

## S7 method for class <kultarr::anchors>
precision(x, model, samples)
}
\arguments{
\item{x}{An object.}

\item{model}{A predict function that returns predicted labels given a
dataset.}

\item{samples}{The dataset to test precision on.}
}
\value{
A named vector of proportions for each class predicted by
\code{model}.
}
\description{
\code{precision()} is an S7 generic that calculates the precision of an
anchor — the proportion of each predicted class among observations
satisfying the anchor — with methods available for the following
classes:
'
}
