% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision.R
\name{precision}
\alias{precision}
\title{Generic function to calculate precision of anchors}
\arguments{
\item{samples}{the dataset to test coverage on}

\item{model}{a predict function that will provide the predicted labels given a dataset}
}
\value{
a named vector of proportions for each class type predicted by the model
}
\description{
Generic function to calculate precision of anchors
}
