% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicate.R
\name{predicate}
\alias{predicate}
\title{Defines a predicate object}
\usage{
predicate(
  feature = character(0),
  operator = function() NULL,
  constant = integer(0)
)
}
\arguments{
\item{feature}{Character. Defines the column that the predicate is applied to}

\item{operator}{Binary operator that works with the column and the constant given}

\item{constant}{Value to compare \code{feature} with using \code{operator}. Can be numeric/character/logical}
}
\value{
A predicate object containing the properties \code{feature}, \code{operator}, \code{constant}
}
\description{
A predicate is built up of a feature column name, a comparison operator and a constant to compare with
For example a predicate can take the form of x > 10, which captures all the observations that have the feature x greater than 10
}
