% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satisfies.R
\name{satisfies}
\alias{satisfies}
\title{Generic function to calculate the number of rows in a dataset satisfied by an anchor}
\arguments{
\item{x}{Object of S7 class anchor}

\item{data}{The dataframe to apply anchors on. Can be one instance or an entire dataset.}
}
\value{
A logical vector indicating whether the anchors satisfies \code{data}
}
\description{
Generic function to calculate the number of rows in a dataset satisfied by an anchor
}
