% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mab_utils.R
\name{validate_bound}
\alias{validate_bound}
\title{Function to validate if a created boundary satisfies a given instance row}
\usage{
validate_bound(b, instance_row)
}
\description{
Function to validate if a created boundary satisfies a given instance row
}
