% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twod_vis.R
\name{vis_anchor}
\alias{vis_anchor}
\title{Visualize the anchor in two dimensional space}
\usage{
vis_anchor(anchors, dataset, instance, model_func)
}
\arguments{
\item{dataset}{The dataset passed to \code{make_anchors()}}

\item{instance}{The point of interest}

\item{model_func}{A crate object containing the prediction function of the modelfor visualisation}

\item{anchor}{The result of \code{make_anchors()} function call.}
}
\value{
A ggplot object
}
\description{
Visualize the anchor in two dimensional space
}
